Fixed nice_load so that it works as expected.
Reworked nice_load as per SO post on loading into rdb or rdx.
Fixes issue #6
Includes rename option to allow renaming.
Also altered travis.yml so that warnings no longer fail build.
This commit depends on exporting a function with `:::` which causes a warning in build, which fails travis.
Can't see a way around the use of `:::`