Fixed bug in rate_confint
rate_confint returned an absolute value for rate differences.
This is not what is desired when more than one level is being compared
and rates in the comparator level may be lower than in the reference
level.

This has been fixed, and the function now returns negative rate
differences if they occur.